moved medical cert to eligible criteria folder
