+++ openfisca_canada/variables/EBF/eligibility_criteria/person_medical_cert.py
@@ -1,0 +1,12 @@
+from openfisca_core.model_api import * 
+# Import the entities specifically defined for this tax and benefit system 
+from openfisca_canada.entities import Person, Family  
+
+
+
+class has_medical_certificate(Variable): 
+    value_type = bool 
+    entity = Person 
+    label = u"does this person have a medical certificate?" 
+    definition_period = DAY 
+    reference = u"TODO"
